+++ docs/source/whatsnew.rst
@@ -29,6 +29,8 @@ Legend
 
 Latest
 ======
+- |Add| extraction of journal and article topics for :code:`arxiv` papers
+  through CLI command :code:`bbs_database topic-extract arxiv`.
 - |Add| extraction of journal and article topics for :code:`pubmed` papers
   through CLI command :code:`bbs_database topic-extract pubmed`.
 - |Add| extraction of journal topics for :code:`pmc` papers through CLI command